use 64 bit ints all around to make things fair.

+++ src/google_dense_hash_map.cc
@@ -1,5 +1,6 @@
+#include <inttypes.h>
 #include <google/dense_hash_map>
-typedef google::dense_hash_map<int, int> hash_t;
+typedef google::dense_hash_map<int64_t, int64_t> hash_t;
 #define SETUP hash_t hash; hash.set_empty_key(-1);
 #define INSERT_INTO_HASH(key, value) hash.insert(hash_t::value_type(key, value))
 #include "template.c"
