+++ src/qt_qhash.cc
@@ -1,5 +1,6 @@
+#include <inttypes.h>
 #include <QHash>
-typedef QHash<int, int> hash_t;
+typedef QHash<int64_t, int64_t> hash_t;
 #define SETUP hash_t hash;
 #define INSERT_INTO_HASH(key, value) hash.insert(key, value)
 #include "template.c"
